Move Joda classes to separate module and create one for MomentJS
