+++ time-lib/joda/src/main/scala/cron4s/lib/joda/JodaInstances.scala
@@ -16,8 +16,8 @@
 
 package cron4s.lib.joda
 
-import cron4s.CronField
-import cron4s.datetime.{DateTimeUnit, IsDateTime}
+import cron4s._
+import cron4s.datetime.{IsDateTime, DateTimeUnit}
 
 import org.joda.time._
 import org.joda.time.base.BaseLocal
